pyproject.toml
[tool.poetry]
name = "kube-web-view"
version = "23.8.0"
description = "Kubernetes Web View allows to list and view all Kubernetes resources (incl. CRDs) with permalink-friendly URLs in a plain-HTML frontend"
authors = ["Henning Jacobs <henning@zalando.de>"]
license = "GPL-3.0+"
readme = "README.md"
homepage = "https://kube-web-view.readthedocs.io/"
documentation = "https://kube-web-view.readthedocs.io/"
repository = "https://codeberg.org/hjacobs/kube-web-view"
packages = [
    { include = "kube_web" }
]

[tool.poetry.scripts]
kube-web-view = 'kube_web:main.main'

[tool.poetry.dependencies]
python = "^3.10"
pykube-ng = ">=19.9.2"
Jinja2 = "^3.0"
aiohttp-jinja2 = "^1.1"
Pygments = "^2.4"
aiohttp_session = {version = "^2.7", extras = ["secure"]}
aioauth-client = "^0.17.3"
aiohttp_remotes = "^0.1.2"
jmespath = "^0.9.4"
[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^3.0"
Sphinx = "^2.1"
sphinx-rtd-theme = "^0.4.3"
requests-html = "^0.10.0"
pytest-kind = ">=22.11.1"
pre-commit = "^2.20"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
